支持node编译quickjs的switch版本。(ref https://github.com/Tencent/puerts/issues/1796 Author:NiceTry12138)
