+++ unity/cli/backends.json
@@ -50,6 +50,11 @@
                     "x64": [
                         "/Lib/Linux/libquickjs.a"
                     ]
+                },
+                "ns":{
+                    "arm64": [
+                        "/Lib/ns/arm64-v8a/libquickjs.a"
+                    ]
                 }
             },
             "copy-libraries": {
@@ -323,6 +328,11 @@
                     ],
                     "arm64": [
                         "/Lib/Linux_arm64/libwee8.a"
+                    ]
+                },
+                "ns":{
+                    "arm64": [
+                        "/Lib/ns/arm64-v8a/libwee8.a"
                     ]
                 }
             },
