+++ unity/cli/cmd.mts
@@ -3,15 +3,14 @@ import { Option, program } from "commander";
 import { join } from "path";
 import downloadBackend from "./backend.mjs";
 import { dotnetTest, unityTest } from "./test.mjs";
-import runPuertsMake from "./make.mjs";
-import { makeOSXUniveralBinary, platformCompileConfig } from "./make.mjs";
+import runPuertsMake, { makeOSXUniveralBinary, platformCompileConfig } from "./make.mjs";
 
 setWinCMDEncodingToUTF8();
 
 const nodePlatformToPuerPlatform = {
     "darwin": "osx",
     "win32": "win"
-};
+}
 const cwd = process.cwd();
 
 program
@@ -19,7 +18,7 @@ program
     .addOption(
         new Option("--platform <platform>", "the target platform")
             .default("")
-            .choices(["win", "osx", "linux", "android", "ios", "ohos", "wasm"])
+            .choices(["win", "osx", "linux", "android", "ios", "ohos", "wasm", "ns"])
     )
     .addOption(
         new Option("--arch <arch>", "the target architecture. 'auto' means build all available archs for the platform and universal binary will be created in osx.")
@@ -34,8 +33,6 @@ program
     .option("--backend <backend>", "the JS backend will be used", "v8_9.4")
     .option('-ws, --websocket <number>', 'with websocket support')
     .option('-G, --generator <generator-name>', 'cmake generator name')
-    .option('-ts, --thread_safe', 'thread safe')
-    .option('-wi, --with_inspector', 'with inspector')
     .action(function (quickcommand, options) {
         let backend = options.backend;
         let config = options.config;
@@ -102,14 +99,15 @@ program
 
         if (options.platform && options.arch == 'auto') {
             let promiseChain = Promise.resolve();
-            const outputs = [];
-            Object.keys(platformCompileConfig[options.platform]).forEach(arch => {
+            const outputs: string[][] = []
+            Object.keys((platformCompileConfig as any)[options.platform]).forEach(arch => {
                 promiseChain = promiseChain.then(function () {
+                    //@ts-ignore
                     options.arch = arch;
                     return runPuertsMake(cwd, options).then(res => {
                         outputs.push(res);
-                    });
-                });
+                    })
+                })
             });
             if (options.platform == 'osx') {
                 // if arch is not specified, make universal binary
@@ -117,37 +115,37 @@ program
             }
 
         } else if (!options.platform && options.arch == 'auto') {
-            options.platform = nodePlatformToPuerPlatform[process.platform];
+            options.platform = (nodePlatformToPuerPlatform as any)[process.platform]
+            //@ts-ignore
             options.arch = process.arch;
             runPuertsMake(cwd, options);
 
         } else {
             runPuertsMake(cwd, options);
         }
-    });
+    })
 
 const backendProgram = program
     .command("backend");
 
 backendProgram
     .command("download [backend] [url]").action((backend, url) => {
-        downloadBackend(cwd, backend, url);
+        downloadBackend(cwd, backend, url)
     });
 backendProgram
     .command("clean").action(() => {
-        rm('-rf', join(cwd, ".backends"));
+        rm('-rf', join(cwd, ".backends"))
     });
 
 program
     .command("dotnet-test [backend]")
     .option("--filter <filter>", "testcase will be filtered", "")
     .option('-sq, --switch_qjs', 'switch to quickjs backend')
-    .option('-ts, --thread_safe', 'thread safe')
-    .action((backend, options) => {
+    .action((backend: string, options: any) => {
         if (options.switch_qjs) {
             process.env.SwitchToQJS = '1';
         }
-        dotnetTest(cwd, backend || "quickjs", options.filter, options.thread_safe);
+        dotnetTest(cwd, backend || "quickjs", options.filter);
     });
 
 program
@@ -155,6 +153,6 @@ program
     .requiredOption("--unity <pathToUnity>")
     .action((options) => {
         unityTest(cwd, options.unity);
-    });
+    })
 
 program.parse(process.argv);
