+++ unity/native_src/CMakeLists.txt
@@ -153,9 +153,15 @@ else ()
             ${PUERTS_SRC} ${PUERTS_INC}
         )
     else ()
-        add_library(puerts SHARED
-            ${PUERTS_SRC} ${PUERTS_INC}
-        )
+        if (SWITCH_PLATFORM) 
+            add_library(puerts STATIC
+                ${PUERTS_SRC} ${PUERTS_INC}
+            )
+        else()
+            add_library(puerts SHARED
+                ${PUERTS_SRC} ${PUERTS_INC}
+            )
+        endif ()
    endif ()
 endif ()
 
@@ -411,6 +417,38 @@ elseif (UNIX)
         set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -stdlib=libc++ -lc++abi")
     endif ()
 endif ()
+
+# ref https://github.com/Tencent/puerts/issues/1796 Author:NiceTry12138 
+if (SWITCH_PLATFORM) 
+    set(NINTENDO_SDK_ROOT $ENV{NINTENDO_SDK_ROOT})
+    set(NS_INCLUDE "${NINTENDO_SDK_ROOT}/Include")
+    set(NS_SPECIFIC_INCLUDE "${NINTENDO_SDK_ROOT}/TargetSpecificInclude/NX-Win32-v142")
+    set(NS_AARCH64 "${NINTENDO_SDK_ROOT}/Compilers/NX/nx/aarch64")
+    
+    set(NS_INCLUDE_LIST ${NS_INCLUDE})                                      # 定义搜索路径
+    list(APPEND NS_INCLUDE_LIST ${NS_SPECIFIC_INCLUDE})                     # 添加特殊文件搜索路径
+    #list(APPEND NS_INCLUDE_LIST "${NS_AARCH64}/include")                    # 添加 AARCH64 头文件
+    list(APPEND NS_INCLUDE_LIST "${NINTENDO_SDK_ROOT}/Include/nn/socket")   # 网络库相关
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fPIC")
+
+    #set(NS_LIB_LIST "$${NS_AARCH64}/lib")                                   # 添加 AARCH64 静态库
+   
+    if(NS_US_ELF)
+        #list(APPEND NS_INCLUDE_LIST "${NS_AARCH64}/include/aarch64-nintendo-nx-elf")
+        #list(APPEND NS_LIB_LIST "${NS_AARCH64}/lib/aarch64-nintendo-nx-elf")
+    else ()
+        #list(APPEND NS_INCLUDE_LIST "${NS_AARCH64}/include/aarch64-nintendo-nx-nncfi")
+        #list(APPEND NS_LIB_LIST "${NS_AARCH64}/lib/aarch64-nintendo-nx-nncfi")
+    endif()
+    
+    add_compile_options(-fdeclspec)
+    add_definitions(-DNN_SDK_BUILD_RELEASE)
+    #add_definitions(-DNN_SDK_BUILD_DEBUG)
+
+    
+    include_directories(${NS_INCLUDE_LIST})
+    link_directories(${NS_LIB_LIST})
+endif()
 
 # link
 target_link_libraries(puerts
